refactored scripts
