--- scripts/helpers/airdrop_data/airdrop_recipients.json
@@ -1,11 +1,0 @@
-{
-  "columns": ["User Address", "Airdrop Amount"],
-  "index": [1, 2, 3, 4, 5],
-  "data": [
-    ["terra17mpuq65hw5kt7d44kpw4nk7x339xhznaa0duzv", 2498087000],
-    ["terra1lv845g7szf9m3082qn3eehv9ewkjjr2kdyz0t6", 2498730003],
-    ["terra1mthww38ea56sjmwtswlvy0g4zzjspjwaw3e6t8", 2503467000],
-    ["terra1jdd392vxlx5u23yvekvvjzuna0gszttgtngfnt", 2503467000],
-    ["terra1yskm9s4r0h0egg3lxe5wmmppr9s6lfau4j8yhc", 2503516700]
-  ]
-}
